Fractional Regression Models

Provides tools for the estimation and specification analysis of both one-part and two-part fractional regression models. It supports various link functions, allows for the calculation of average and conditional partial effects, and provides several specification tests such as RESET, P test, and GGOFF tests.

Details

Package:	fracreg

Type:	Package

Version:	1.0.0

Date:	2026-07-05

License:	GPL-3

Author

Sulman Olieko Owili <oliekosulman@gmail.com>

References

Ramalho, E.A., J.J.S. Ramalho and J.M.R. Murteira (2011), "Alternative estimating and testing empirical strategies for fractional regression models", Journal of Economic Surveys, 25(1), 19-68.